--- env.js
@@ -1,6 +1,0 @@
-module.exports = {
-    devServer: {
-        host: 'localhost',
-        port: 3000
-    }
-}
