+++ package.json
@@ -1,45 +1,53 @@
 {
   "name": "facebook.clone",
-  "version": "1.0.0",
-  "description": "A Facebook clone built with React and material.ui",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "webpack-dev-server": "webpack-dev-server",
-    "dev": "webpack serve --open --config webpack/webpack.config.js --mode=development",
-    "prod": "webpack --mode=production"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/konichar/facebook.clone.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/konichar/facebook.clone/issues"
-  },
-  "homepage": "https://github.com/konichar/facebook.clone#readme",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "file-loader": "^6.2.0",
-    "prop-types": "^15.7.2",
+    "@snowpack/app-scripts-react": "^2.0.1",
+    "@snowpack/plugin-babel": "^2.1.7",
+    "@snowpack/plugin-build-script": "^2.1.0",
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "postcss-cli": "^8.3.1",
+    "postcss-flexbugs-fixes": "^5.0.2",
+    "postcss-import": "^14.0.1",
+    "postcss-normalize": "^9.0.0",
+    "postcss-preset-env": "^6.7.0",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "react-scripts": "4.0.3",
+    "rollup-plugin-node-polyfills": "^0.2.1",
+    "web-vitals": "^1.0.1"
+  },
+  "scripts": {
+    "start": "snowpack dev --config snowpack.config.js",
+    "start:cra": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   },
   "devDependencies": {
-    "@babel/core": "^7.13.14",
-    "@babel/plugin-proposal-class-properties": "^7.13.0",
-    "@babel/preset-env": "^7.13.12",
-    "@babel/preset-react": "^7.13.13",
-    "babel-loader": "^8.2.2",
-    "babel-preset-react": "^6.24.1",
-    "css-loader": "^5.2.0",
-    "html-webpack-plugin": "^5.3.1",
-    "style-loader": "^2.0.0",
-    "webpack": "^5.30.0",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2"
+    "snowpack": "^3.2.2"
   }
 }
