+++ public/index.js
@@ -1,7 +1,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import App from '../src/app';
 
 ReactDOM.render(
   <React.StrictMode>
