+++ src/components/profile/profile.js
@@ -1,14 +1,16 @@
 import React from 'react'
-import Header from './profile-header'
-import ProfileBody from './profile-body'
+import ProfileHeader from './ProfileHeader'
+import ProfileBody from './ProfileBody'
 
 const Profile = () => {
+  const [tabvalue, setTabValue] = React.useState(0);
+  const handleChange = (event, newValue) => {
+    setTabValue(newValue);
+  };
     return(
         <main>
-            {/* <Header/> */}
-            <ProfileBody>
-
-            </ProfileBody>
+            <ProfileHeader value={tabvalue} handleChange={handleChange}/>
+            <ProfileBody value={tabvalue} handleChange={handleChange}/>
         </main>
     )
 }
