+++ src/components/subcomponents/ProfileAbout.jsx
@@ -1,0 +1,8 @@
+
+const ProfileAbout = () => {
+    return (
+        <div>this is about</div>
+    )
+}
+
+export default ProfileAbout
