+++ src/components/subcomponents/ProfilePost.jsx
@@ -1,0 +1,38 @@
+import {
+  Box,
+  Grid,
+  GridList,
+  GridListTile,
+  GridListTileBar,
+} from "@material-ui/core";
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    display: "flex",
+    flexWrap: "wrap",
+    justifyContent: "space-around",
+    overflow: "hidden",
+    backgroundColor: theme.palette.background.paper,
+  },
+  gridList: {
+    width: 500,
+    height: 450,
+    // Promote the list into his own layer on Chrome. This cost memory but helps keeping high FPS.
+    transform: "translateZ(0)",
+  },
+}));
+
+const ProfilePost = () => {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.root}>
+      <Grid>
+        
+      </Grid>
+    </div>
+  );
+};
+
+export default ProfilePost;
